move getDOMInfo to core
