+++ packages/core/package.json
@@ -36,8 +36,10 @@
   "dependencies": {
     "@craftjs/utils": "^0.2.1",
     "debounce": "^2.0.0",
+    "immer": "^9.0.6",
     "lodash": "^4.17.21",
     "nanoid": "^3.1.23",
+    "shallowequal": "^1.1.0",
     "tiny-invariant": "^1.0.6"
   },
   "devDependencies": {
