+++ packages/core/src/utils/getDOMInfo.ts
